Asignando funciones a puntos, recompensas, reciclaje y usuario

+++ reusappbackend/src/main/kotlin/com/example/models/Reciclaje.kt
@@ -6,9 +6,9 @@ import kotlinx.serialization.Serializable
 import java.time.LocalDate
 
 @Serializable
-class Reciclaje (
+class Reciclaje(
     val material: TipoMaterial,
-    val peso: Float,
+    val peso: Double,
     @Contextual
     val fecha: LocalDate,
     var usuario: Usuario,
